Refactor for DRY
